Remove ca-certificates-java
Its circular dependency with the JDK packages makes the AdoptOpenJDK packages
uninstallable on Debian. Trying to somehow get it to install anyway would result
in brittle packages because apt/dpkg provides no guarantees how dependency
cycles are going to be resolved and it’s largely configuration-dependant.

Without ca-certificates-java, the JDK tools will use the default truststore
shipped with the JDK.