Separate RPM prerm scripts of JDK/JRE (#230)
Previously, both JDK/JRE shared the same prerm script for RPMs. The script tried to remove the alternative for javac which isn't part of the JRE. Consequently, the script failed and so failed the JRE RPMs to uninstall. With this change, the JDK and JRE have separate prerm scripts and the one for the JRE does not try to remove the alternative for the non-existent javac.

Fixes #215.