Move quotes around SIGN_CERT into variable itself
When CERTIFICATE is not used, the empty quotes around
SIGN_CERT in the packagebuild command will mess up the
input args. Moving the quotes into the variable will
remove them from the call when the variable doesn't get set.
Also adjust the IF to explicitly set the sign variables
to blank to avoid unintended behaviour.

Fixes #256
Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>