Separate RPM generation for SUSE from Red Hat (#285)
So far, we were able to generate SUSE and Red Hat RPMs from the same
configuration (with some cheating around paths). This is no longer tenable
because also the list of dependencies started to diverge. For example,
the DejaVu fonts are called dejavu-sans-fonts in Red Hat flavours while SUSE
calls them dejavu-fonts.